Replace `Clone` impl. for `NonFinalizedState`
It was previously derived, but since the state will have a
`tokio::sync::watch::Sender` field, which is not cloneable, it can't be
cloned anymore. Unfortunately, just removing the derived implementation
breaks some tests, so a manual implementation of `Clone` was introduced
that's only available for tests. The watch channel will then be ignored
when cloning.